backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
supabase==2.4.5
marshmallow==3.20.1
bcrypt==4.1.2
python-jose==3.3.0
gunicorn==21.2.0
requests==2.31.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
razorpay==1.4.1
werkzeug==3.0.1
httpx==0.27.2
postgrest==0.16.2
Flask-Limiter==3.5.0
Flask-Compress==1.14
bleach==6.1.0
gevent==24.2.1
waitress==3.0.0
google-api-python-client==2.115.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
